Extract run command from cli/index.ts
