+++ cli/commands/index.ts
@@ -4,4 +4,5 @@ export { helpCommand } from "df/cli/commands/help_command";
 export { initCommand } from "df/cli/commands/init_command";
 export { initCredsCommand } from "df/cli/commands/init_creds_command";
 export { installCommand } from "df/cli/commands/install_command";
+export { runCommand } from "df/cli/commands/run_command";
 export { testCommand } from "df/cli/commands/test_command";
